# projects/01-ai-knowledge-base/docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.CPUMetrics:
    properties:
      cores:
        type: integer
      usage_percent:
        type: number
    type: object
  handler.DiskMetrics:
    properties:
      free:
        type: integer
      path:
        type: string
      total:
        type: integer
      used:
        type: integer
      used_percent:
        type: number
    type: object
...
  handler.MemoryMetrics:
    properties:
      free:
        type: integer
      total:
        type: integer
      used:
        type: integer
      used_percent:
        type: number
    type: object
  handler.NetworkMetrics:
    properties:
      bytes_recv:
        type: integer
      bytes_sent:
        type: integer
    type: object
  handler.RuntimeMetrics:
    properties:
      go_version:
        type: string
      goroutines:
        type: integer
      last_gc_time:
        type: string
      mem_alloc_mb:
        type: integer
      mem_total_mb:
        type: integer
      num_cpu:
        type: integer
      num_gc:
        type: integer
    type: object
...
  handler.ServiceInfo:
    properties:
      available:
        type: boolean
      latency:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
  handler.SystemMetrics:
    properties:
      cpu:
        $ref: '#/definitions/handler.CPUMetrics'
      disk:
        $ref: '#/definitions/handler.DiskMetrics'
      memory:
        $ref: '#/definitions/handler.MemoryMetrics'
      network:
        $ref: '#/definitions/handler.NetworkMetrics'
    type: object
  handler.SystemStatusResponse:
    properties:
      runtime:
        $ref: '#/definitions/handler.RuntimeMetrics'
      services:
        additionalProperties:
          $ref: '#/definitions/handler.ServiceInfo'
        type: object
      status:
        type: string
      system:
        $ref: '#/definitions/handler.SystemMetrics'
      timestamp:
        type: string
      uptime:
        type: string
      version:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@saleschampionhub.com
    name: 技术支持
  description: "企业级AI知识库管理平台，支持多租户、RLS数据隔离和向量搜索功能。\n\n## \U0001F510 认证说明\n所有接口都需要JWT
    Token认证（从子项目0用户中心获取）。\n使用用户中心的 `/api/v1/auth/login` 接口获取Token。\n\n## ✨ 核心特性\n-
    多租户架构：租户/组织/用户三级隔离\n- Row-Level Security (RLS)：数据库级别的数据隔离\n- 三级挂载系统：租户 > 组织 >
    用户权限继承\n- 向量搜索：基于pgvector的语义搜索（1024维向量）\n- Redis多层缓存：热点数据缓存优化\n- 完整用户中心集成：统一认证授权"
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: AI知识库管理平台 API
  version: "1.0"
paths:
...
  /api/v1/system/status:
    get:
      description: 获取详细的系统状态信息，包括CPU、内存、磁盘、网络等指标
      produces:
      - application/json
      responses:
        "200":
          description: 系统状态信息
          schema:
            $ref: '#/definitions/handler.SystemStatusResponse'
      security:
      - BearerAuth: []
      summary: 获取系统状态
      tags:
      - 监控
...
securityDefinitions:
  BearerAuth:
    description: 输入格式："Bearer " + JWT令牌（注意Bearer后面有空格）
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"